move descriptions from api to lib
