+++ api/types/usertasks/object.go
@@ -19,9 +19,7 @@
 package usertasks
 
 import (
-	"embed"
 	"encoding/binary"
-	"fmt"
 	"slices"
 	"strconv"
 	"time"
@@ -155,21 +153,6 @@ const (
 	// This happens when there's a failure with permissions or an invalid configuration (eg, invalid document name).
 	AutoDiscoverEC2IssueSSMInvocationFailure = "ec2-ssm-invocation-failure"
 )
-
-//go:embed descriptions/*.md
-var descriptionsFS embed.FS
-
-// DescriptionForDiscoverEC2Issue returns the description of the issue and fixing steps.
-// The returned string contains a markdown document.
-// If issue type is not recognized or doesn't have a specific description, them an empty string is returned.
-func DescriptionForDiscoverEC2Issue(issueType string) string {
-	filename := fmt.Sprintf("descriptions/%s.md", issueType)
-	bs, err := descriptionsFS.ReadFile(filename)
-	if err != nil {
-		return ""
-	}
-	return string(bs)
-}
 
 // DiscoverEC2IssueTypes is a list of issue types that can occur when trying to auto enroll EC2 instances.
 var DiscoverEC2IssueTypes = []string{
