+++ api/types/usertasks/object_test.go
@@ -465,9 +465,3 @@ func TestNewDiscoverEKSUserTask(t *testing.T) {
 		})
 	}
 }
-
-func TestAllDescriptions(t *testing.T) {
-	for _, issueType := range usertasks.DiscoverEC2IssueTypes {
-		require.NotEmpty(t, usertasks.DescriptionForDiscoverEC2Issue(issueType), "issue type %q is missing descriptions/%s.md file", issueType, issueType)
-	}
-}
