+++ lib/usertasks/descriptions/ec2-ssm-agent-not-registered.md
